[Feat] 페이징 변환 유틸 추가

+++ src/main/java/kr/aling/admin/common/dto/PageResponseDto.java
@@ -4,12 +4,19 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+/**
+ * 페이징에 필요한 내용을 담는 페이징 응답 Dto.
+ *
+ * @param <T> 페이징 내용 타입
+ * @author : 이수정
+ * @since : 1.0
+ */
 @Getter
 @AllArgsConstructor
 public class PageResponseDto<T> {
 
-    private long pageNumber;
-    private long totalPages;
-    private long totalElements;
+    private Integer pageNumber;
+    private Integer totalPages;
+    private Long totalElements;
     private List<T> content;
 }
